chore: merge duplicate code to share util config route  🎨
